+++ example/lib/main.dart
@@ -1,5 +1,8 @@
+import 'dart:developer';
+
 import 'package:flutter/material.dart';
 import 'package:theme_patrol/theme_patrol.dart';
+import 'package:animated_checkmark/animated_checkmark.dart';
 
 void main() {
   runApp(const MyApp());
@@ -14,13 +17,12 @@ class MyApp extends StatelessWidget {
     return ThemePatrol(
       // initialTheme: 'amber',
       // initialMode: ThemeMode.system,
-      onAvailableChanged: (_) => print('available themes changed'),
-      onThemeChanged: (theme) => print('theme changed to ${theme.selected}'),
-      onModeChanged: (theme) =>
-          print('theme mode changed to ${theme.mode.name}'),
+      onAvailableChanged: (_) => log('available themes changed'),
+      onThemeChanged: (theme) => log('theme changed to ${theme.selected}'),
+      onModeChanged: (theme) => log('theme mode changed to ${theme.mode.name}'),
       onColorChanged: (theme) =>
-          print('theme color changed to ${theme.color.toString()}'),
-      onChanged: (theme) => print('value changed'),
+          log('theme color changed to ${theme.color.toString()}'),
+      onChanged: (theme) => log('value changed'),
       themes: {
         'basic': ThemeConfig.fromColor(Colors.purple),
         'pro': ThemeConfig.fromColor(Colors.red),
@@ -67,7 +69,13 @@ class MyHomePageState extends State<MyHomePage> {
                   children: [
                     ElevatedButton.icon(
                       onPressed: () => theme.toggleMode(),
-                      icon: Icon(theme.modeIcon),
+                      icon: AnimatedSwitcher(
+                        duration: const Duration(milliseconds: 200),
+                        child: Icon(
+                          theme.modeIcon,
+                          key: ValueKey(theme.mode),
+                        ),
+                      ),
                       label: Text('${theme.mode.name.toUpperCase()} MODE'),
                     ),
                     const SizedBox(height: 10),
@@ -127,34 +135,35 @@ class MyHomePageState extends State<MyHomePage> {
               style: TextStyle(fontWeight: FontWeight.w400),
             ),
             const SizedBox(height: 10),
-            Container(
-              width: 200,
-              alignment: Alignment.center,
-              child: GridView.builder(
-                shrinkWrap: true,
-                itemCount: Colors.primaries.length,
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                  mainAxisSpacing: 2,
-                  crossAxisSpacing: 2,
-                  crossAxisCount: 6,
+            ThemeConsumer(builder: (context, theme) {
+              return Container(
+                width: 200,
+                alignment: Alignment.center,
+                child: GridView.builder(
+                  shrinkWrap: true,
+                  itemCount: Colors.primaries.length,
+                  gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+                    mainAxisSpacing: 2,
+                    crossAxisSpacing: 2,
+                    crossAxisCount: 6,
+                  ),
+                  itemBuilder: (_, i) {
+                    final color = Colors.primaries[i];
+                    return Card(
+                      color: color,
+                      child: InkWell(
+                        onTap: () => theme.toColor(color),
+                        child: AnimatedCheckmark(
+                          weight: 4,
+                          color: Colors.white70,
+                          active: color == theme.color,
+                        ),
+                      ),
+                    );
+                  },
                 ),
-                itemBuilder: (_, i) {
-                  final color = Colors.primaries[i];
-                  return Card(
-                    color: color,
-                    child: InkWell(
-                      onTap: () => ThemePatrol.of(context).toColor(color),
-                      child: color == ThemePatrol.of(context).color
-                          ? const Icon(
-                              Icons.check,
-                              color: Colors.white,
-                            )
-                          : Container(),
-                    ),
-                  );
-                },
-              ),
-            ),
+              );
+            }),
             const SizedBox(height: 10),
             TextButton(
               onPressed: () => ThemePatrol.of(context).resetColor(),